[Feature] Support responses returned by before hooks
Allows a response to be returned from any of the controller
before hooks, e.g. reading, saving, deleting etc. If a value is
returned from these hooks, the action will stop and return the
value immediately.

Closes #24

+++ src/Http/Controllers/Actions/Destroy.php
@@ -42,17 +42,20 @@ trait Destroy
         ) ?? \request();
 
         $model = $route->model();
+        $response = null;
 
         if (method_exists($this, 'deleting')) {
-            $this->deleting($model, $request);
+            $response = $this->deleting($model, $request);
+        }
+
+        if ($response) {
+            return $response;
         }
 
         $store->delete(
             $resourceType,
             $route->modelOrResourceId()
         );
-
-        $response = null;
 
         if (method_exists($this, 'deleted')) {
             $response = $this->deleted($model, $request);
